Add lmoments (#54)

+++ tests/test_poisson.py
@@ -17,7 +17,7 @@ from tests.helper_scipy import make_params, run_distribution_tests
 )
 def test_poisson_vs_scipy(params, sp_params):
     """Test Poisson distribution against scipy.stats.poisson."""
-    p_params = make_params(*params)
+    p_params = make_params(*params, dtype="float64")
     support = (0, float("inf"))
 
     run_distribution_tests(
